Update nextjs structure and prep for static
