+++ dapps/react-dapp-v2-next/package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-dapp-v2-next",
-  "version": "0.1.0",
+  "version": "2.0.0-6de2113.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -9,33 +9,42 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@ethereumjs/tx": "^3.5.2",
-    "@solana/web3.js": "^1.50.1",
+    "@ethereumjs/tx": "^3.5.0",
+    "@solana/web3.js": "^1.36.0",
     "@walletconnect/encoding": "^1.0.1",
-    "@walletconnect/qrcode-modal": "^1.8.0",
-    "@walletconnect/types": "^1.8.0",
-    "@walletconnect/utils": "^1.8.0",
-    "axios": "^0.27.2",
+    "@walletconnect/qrcode-modal": "^1.7.8",
+    "@walletconnect/sign-client": "2.0.0-6de2113.0",
+    "@walletconnect/types": "2.0.0-6de2113.0",
+    "@walletconnect/utils": "2.0.0-6de2113.0",
+    "axios": "^0.21.1",
     "blockies-ts": "^1.0.0",
     "bs58": "^5.0.0",
-    "caip-api": "^1.4.6",
+    "caip-api": "^2.0.0-beta.1",
     "cosmos-wallet": "^1.2.0",
-    "eth-sig-util": "^3.0.1",
-    "ethereumjs-util": "^7.1.5",
-    "ethers": "^5.6.9",
+    "eth-sig-util": "^2.5.3",
+    "ethereumjs-util": "^7.0.6",
+    "ethers": "^5.3.0",
     "next": "12.2.4",
-    "prop-types": "^15.8.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "prop-types": "^15.7.2",
+    "qr-image": "^3.2.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-scripts": "^4.0.3",
     "solana-wallet": "^1.0.1",
-    "styled-components": "^5.3.5",
-    "web-vitals": "^2.1.4"
+    "styled-components": "^5.2.0",
+    "web-vitals": "^0.2.4"
   },
   "devDependencies": {
-    "@types/node": "18.6.4",
+    "@types/eth-sig-util": "^2.1.1",
+    "@types/jest": "^27.4.0",
+    "@types/node": "^17.0.14",
+    "@types/pino": "^7.0.5",
+    "@types/prop-types": "^15.7.4",
+    "@types/qr-image": "^3.2.5",
     "@types/react": "18.0.15",
     "@types/react-dom": "18.0.6",
     "@types/styled-components": "^5.1.25",
+    "better-sqlite3": "^7.6.2",
     "eslint": "8.21.0",
     "eslint-config-next": "12.2.4",
     "typescript": "^4.7.4"
