+++ dapps/react-dapp-v2-next/src/components/Column.tsx
