+++ dapps/react-dapp-v2-next/src/constants/logo.ts
