+++ dapps/react-dapp-v2-next/src/contexts/JsonRpcContext.tsx
