+++ dapps/react-dapp-v2-next/src/helpers/eip1271.ts
