+++ dapps/react-dapp-v2-next/src/helpers/namespaces.ts
