+++ dapps/react-dapp-v2-next/src/modals/PingModal.tsx
