+++ dapps/react-dapp-v2-next/src/pages/404.tsx
@@ -1,0 +1,5 @@
+export default function FourOhFour() {
+	return (
+		<h1>404 Page Not Found</h1>
+	)
+}
