+++ dapps/react-dapp-v2-next/src/pages/_app.tsx
