+++ dapps/react-dapp-v2-next/src/pages/_error.tsx
@@ -1,0 +1,7 @@
+export default function Error() {
+	return (
+		<div>
+			An error as occured
+		</div>
+	)
+}
