+++ dapps/react-dapp-v2-next/src/styles/globals.css
